requirements.txt
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
ffmpeg-python==0.2.0
filelock==3.10.7
future==0.18.3
huggingface-hub==0.13.3
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.0
numba==0.56.4
numpy==1.23.5
openai-whisper==20230308
packaging==23.0
pathy==0.10.1
preshed==3.0.8
psutil==5.9.4
pydantic==1.10.7
pydub==0.25.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
smart-open==6.3.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
thinc==8.1.9
tokenizers==0.13.2
torch==2.0.0
tqdm==4.65.0
transformers==4.27.4
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
